feat(components): move LinkExternal to shared component, import into admin
