+++ packages/fxa-components/LinkExternal/index.tsx
@@ -2,7 +2,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-import React from 'react';
+import React from "react";
 
 type LinkExternalProps = {
   className?: string;
@@ -13,7 +13,7 @@ type LinkExternalProps = {
 export const LinkExternal = ({
   className,
   href,
-  children,
+  children
 }: LinkExternalProps) => (
   <a target="_blank" rel="noopener noreferrer" {...{ className }} {...{ href }}>
     {children}
